docs: initialize memfuse central audit report and review plan
Confirm agreement with the iterative 3-phase audit strategy across all 15
MemFuse workspace crates and initialize `memfuse_audit_report.md` with
detailed focus areas (SIMD/Mmap safety, RRF fusion tie-breaking, LSM/WAL
fsync propagation, 2PC compensation, PyO3 Tokio isolation, stdio JSON-RPC).

+++ crates/memfuse-router/src/router.rs
@@ -141,6 +141,7 @@ impl RouterEngine {
         }
 
         // 1. Perform hybrid search with standard fusion weights
+        #[allow(deprecated)]
         let search_results = self
             .collection
             .hybrid_search_with_strategy(query_text, query_embedding, 10, None, None, None, None)
@@ -180,7 +181,7 @@ impl RouterEngine {
                 .map(|p| {
                     let mut ep = p.clone();
                     if let Some(state) = cal.get(&p.name) {
-                        if state.conformal.window_total >= CALIBRATION_WARMUP_WINDOW as u64 {
+                        if state.conformal.window_total >= CALIBRATION_WARMUP_WINDOW {
                             ep.min_relevance_score = state.calibrated_min_score;
                         }
                     }
